training/runtime-requirements.txt
# Generated by KFP.
google-cloud-aiplatform==1.25.0
numpy
pandas
tensorflow